Make the pipeline stages attend to url changes
